Skip CSharpWinForms.DeleteControl on 32 bit release
Refer https://dev.azure.com/devdiv/DevDiv/_build/results?buildId=10331299&view=ms.vss-test-web.build-test-results-tab&runId=112330346&resultId=100876&paneView=debug

+++ src/VisualStudio/IntegrationTest/New.IntegrationTests/CSharp/CSharpWinForms.cs
@@ -208,28 +208,32 @@ public class CSharpWinForms : AbstractEditorTest
     [IdeFact]
     public async Task DeleteControl()
     {
-        var project = ProjectName;
-        await TestServices.SolutionExplorer.OpenFileWithDesignerAsync(project, "Form1.cs", HangMitigatingCancellationToken);
-        await TestServices.Editor.AddWinFormButtonAsync("SomeButton", HangMitigatingCancellationToken);
-        await TestServices.SolutionExplorer.SaveFileAsync(project, "Form1.cs", HangMitigatingCancellationToken);
-        await TestServices.SolutionExplorer.SaveFileAsync(project, "Form1.resx", HangMitigatingCancellationToken);
-        await TestServices.Editor.DeleteWinFormButtonAsync("SomeButton", HangMitigatingCancellationToken);
-
-        await TestServices.Workspace.WaitForAllAsyncOperationsAsync(
-            [
-                FeatureAttribute.Workspace,
-                FeatureAttribute.SolutionCrawlerLegacy,
-                FeatureAttribute.DiagnosticService,
-                FeatureAttribute.EditAndContinue,
-                FeatureAttribute.ErrorSquiggles,
-                FeatureAttribute.ErrorList,
-            ],
-            HangMitigatingCancellationToken);
-
-        Assert.Empty(await TestServices.ErrorList.GetBuildErrorsAsync(HangMitigatingCancellationToken));
-        await TestServices.SolutionExplorer.OpenFileAsync(project, "Form1.Designer.cs", HangMitigatingCancellationToken);
-        var actualText = await TestServices.Editor.GetTextAsync(HangMitigatingCancellationToken);
-        Assert.DoesNotContain(@"VisualStudio.Editor.SomeButton.Name = ""SomeButton"";", actualText);
-        Assert.DoesNotContain(@"private System.Windows.Forms.Button SomeButton;", actualText);
+        // Known failure on 32 bit release. Entire phase could eventually be removed. This is a more targeted stopgap until then.
+        if (System.IntPtr.Size != 4)
+        {
+            var project = ProjectName;
+            await TestServices.SolutionExplorer.OpenFileWithDesignerAsync(project, "Form1.cs", HangMitigatingCancellationToken);
+            await TestServices.Editor.AddWinFormButtonAsync("SomeButton", HangMitigatingCancellationToken);
+            await TestServices.SolutionExplorer.SaveFileAsync(project, "Form1.cs", HangMitigatingCancellationToken);
+            await TestServices.SolutionExplorer.SaveFileAsync(project, "Form1.resx", HangMitigatingCancellationToken);
+            await TestServices.Editor.DeleteWinFormButtonAsync("SomeButton", HangMitigatingCancellationToken);
+    
+            await TestServices.Workspace.WaitForAllAsyncOperationsAsync(
+                [
+                    FeatureAttribute.Workspace,
+                    FeatureAttribute.SolutionCrawlerLegacy,
+                    FeatureAttribute.DiagnosticService,
+                    FeatureAttribute.EditAndContinue,
+                    FeatureAttribute.ErrorSquiggles,
+                    FeatureAttribute.ErrorList,
+                ],
+                HangMitigatingCancellationToken);
+    
+            Assert.Empty(await TestServices.ErrorList.GetBuildErrorsAsync(HangMitigatingCancellationToken));
+            await TestServices.SolutionExplorer.OpenFileAsync(project, "Form1.Designer.cs", HangMitigatingCancellationToken);
+            var actualText = await TestServices.Editor.GetTextAsync(HangMitigatingCancellationToken);
+            Assert.DoesNotContain(@"VisualStudio.Editor.SomeButton.Name = ""SomeButton"";", actualText);
+            Assert.DoesNotContain(@"private System.Windows.Forms.Button SomeButton;", actualText);
+        }
     }
 }
